// src/__generated__/leaveTeam.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: leaveTeam
// ====================================================

export interface leaveTeam_leaveTeam {
  __typename: "LeaveTeamOutput";
  ok: boolean;
  error: string | null;
}

export interface leaveTeam {
  leaveTeam: leaveTeam_leaveTeam;
}